Added config code for db connection

+++ src/main/kotlin/com/example/routes/CustomerRoutes.kt
@@ -23,7 +23,7 @@ fun Route.customerRouting() {
                 status = HttpStatusCode.BadRequest
             )
 
-            val customer = customerStorage.find { it.id == id } ?: return@get call.respondText(
+            val customer = customerStorage.find { it.id == id.toInt() } ?: return@get call.respondText(
                 "No customer with id: $id",
                 status = HttpStatusCode.NotFound
             )
@@ -37,7 +37,7 @@ fun Route.customerRouting() {
         }
         delete("{id}") {
             val id = call.parameters["id"] ?: return@delete call.respond(HttpStatusCode.BadRequest)
-            if (customerStorage.removeIf{it.id == id}) {
+            if (customerStorage.removeIf{it.id == id.toInt()}) {
                 call.respondText("Customer Removed Correctly", status = HttpStatusCode.Accepted)
             } else {
                 call.respondText("Not Found", status = HttpStatusCode.NotFound)
